requirements.txt
Pillow == 10.0.1
requests == 2.31.0
build == 1.0.3
twine == 4.0.2
Sphinx == 7.2.6